requirements.txt
langchain-core~=0.3.15
pandas~=2.2.2
python-dotenv~=1.0.1
openai~=1.61.1
openpyxl~=3.1.5
promptflow>=1.8.0
tenacity~=9.0.0
diskcache~=5.6.3
azure-ai-inference~=1.0.0b8
aiohttp~=3.11.11
